requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.50.0
colorama>=0.4.4
numpy>=1.19.0